+++ deps/miniupnpc/minisoap.c
@@ -50,7 +50,7 @@ httpWrite(int fd, const char * body, int bodysize,
 	memcpy(p, headers, headerssize);
 	memcpy(p+headerssize, body, bodysize);
 	/*n = write(fd, p, headerssize+bodysize);*/
-	n = send(fd, p, headerssize+bodysize, 0);
+	n = (int)send(fd, p, headerssize+bodysize, 0);
 	if(n<0) {
 	  PRINT_SOCKET_ERROR("send");
 	}
